requirements.txt
pinecone_client==2.2.2
python-dotenv==1.0.0
streamlit_chat==0.1.1
